Refs #3661. Better filtering for container_guid in pages.

+++ mod/pages/pages/pages/edit.php
@@ -7,15 +7,15 @@
 
 gatekeeper();
 
-$page_guid = get_input('guid');
+$page_guid = (int)get_input('guid');
 $page = get_entity($page_guid);
 if (!$page) {
-
+	
 }
 
 $container = $page->getContainerEntity();
 if (!$container) {
-
+	
 }
 
 elgg_set_page_owner_guid($container->getGUID());
